+++ repentogon/LuaInterfaces/Room/LuaRoom.cpp
@@ -388,7 +388,7 @@ LUA_FUNCTION(Lua_RoomSetItemPool) {
 	Room* room = lua::GetUserdata<Room*>(L, 1, lua::Metatables::ROOM, lua::metatables::RoomMT);
 	const int poolType = (int)luaL_checkinteger(L, 2);
 
-	if (poolType < POOL_NULL || poolType >= CustomItemPool::itemPools.size() + NUM_ITEMPOOLS) {
+	if (poolType < POOL_NULL || poolType >= (int)CustomItemPool::itemPools.size() + NUM_ITEMPOOLS) {
 		return luaL_argerror(L, 2, "Invalid ItemPoolType");
 	}
